refactor: create a separate component for the discount amount
